requirements.txt
Django==1.9.1
django-widget-tweaks==1.4.1
djangorestframework==3.3.2
pyodbc==3.0.10
pypiwin32==219
wheel==0.24.0
